Refactor and enhance code documentation across multiple modules
- Translated comments and documentation from Chinese to English for better accessibility.
- Improved error handling and code clarity in `session_erasure_builder.rs` and `staging_pool.rs`.
- Simplified `LocalStorageWriter` and `LocalStorageReader` structures by removing unused fields.
- Enhanced validation checks in `config.rs` for erasure coding parameters.
- Updated tests to ensure proper functionality and coverage, including concurrency tests and floating footer recovery.
- Added new utility functions for packing and unpacking files in `packed_chunk.rs`.
- Introduced a new `cat` file for potential future use.

+++ crates/era-common/src/conversion.rs
@@ -152,8 +152,8 @@ impl From<ErasureBlockInfo> for proto::ErasureBlockInfo {
         Self {
             data_shards: info.data_shards as u32,
             parity_shards: info.parity_shards as u32,
-            shard_size: info.shard_size as u32,
-            original_len: info.original_len as u32,
+            shard_size: info.shard_size,
+            original_len: info.original_len,
         }
     }
 }
@@ -164,8 +164,8 @@ impl TryFrom<proto::ErasureBlockInfo> for ErasureBlockInfo {
         Ok(Self {
             data_shards: proto.data_shards as u8,
             parity_shards: proto.parity_shards as u8,
-            shard_size: proto.shard_size as u32,
-            original_len: proto.original_len as u32,
+            shard_size: proto.shard_size,
+            original_len: proto.original_len,
         })
     }
 }
